data_engineering/crawler_supabase_fastAPI/requirements.txt
beautifulsoup4==4.13.3
feedparser==6.0.11
httpcore==1.0.8
httpx==0.28.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
supabase==2.15.0
fastapi
uvicorn
pytz
pandas==2.2.3